Don't start proxy for a subpath of app_direct
This ensures that AJAX requests, assets etc of an app does not cause the
app to be started. This is required for implementing the app stop feature.
Without this fix, the app would re-start if it performs an AJAX request
in the still-existing iframe.
Furthermore the 410 status code and specific response body indicate to
the front-end that the app has (purposely) been stopped.