issues/269 redirect user to first available app when proxy.landing-page=auto
Allows ShinyProxy to use the first available app for a logged-in user as landing page, bypassing the built-in index page. This can be used when different permissions for different users makes it impractical to identify an app which is accessible by all users in the system and usable as predefined landing page/app.